+++ src/RestApi/StoreRestController.php
@@ -2,6 +2,7 @@
 
 namespace Foodsharing\RestApi;
 
+use DateTime;
 use Foodsharing\Lib\Session;
 use Foodsharing\Modules\Bell\BellGateway;
 use Foodsharing\Modules\Bell\DTO\Bell;
@@ -226,7 +227,7 @@ class StoreRestController extends AbstractFOSRestController
 		}
 		$result = $this->storeGateway->getStoreWallpost($storeId, $postId);
 
-		$this->storeGateway->addStoreLog($result['betrieb_id'], $this->session->id(), $result['foodsaver_id'], new \DateTime($result['zeit']), StoreLogAction::DELETED_FROM_WALL, $result['text']);
+		$this->storeGateway->addStoreLog($result['betrieb_id'], $this->session->id(), $result['foodsaver_id'], new DateTime($result['zeit']), StoreLogAction::DELETED_FROM_WALL, $result['text']);
 
 		$this->storeGateway->deleteStoreWallpost($storeId, $postId);
 
@@ -494,6 +495,83 @@ class StoreRestController extends AbstractFOSRestController
 		$this->storeTransactions->moveMemberToRegularTeam($storeId, $userId);
 
 		return $this->handleView($this->view([], 200));
+	}
+
+	/**
+	 * Returns an array of log entries with foodsaver and log information.
+	 * The log contains only entries from the past 7 days.
+	 *
+	 * @OA\Parameter(name="storeId", in="path", @OA\Schema(type="integer"))
+	 * @OA\Parameter(name="storeLogActionIds", in="path", @OA\Schema(type="string"), description="The ids of the actions, seperated by commas like: 1,2,3")
+	 *
+	 * @OA\Tag(name="stores")
+	 *
+	 * @Rest\Get("stores/{storeId}/log/{storeLogActionIds}")
+	 */
+	public function showStoreLogHistoryAction(int $storeId, string $storeLogActionIds): Response
+	{
+		if (!$this->session->id()) {
+			throw new UnauthorizedHttpException('', self::NOT_LOGGED_IN);
+		}
+
+		if (!$this->storePermissions->maySeePickupHistory($storeId)) {
+			throw new AccessDeniedHttpException();
+		}
+
+		$storeLogActions = explode(',', $storeLogActionIds);
+		$storeLogEntries = $this->storeGateway->getStoreLogsByActionType($storeId, $storeLogActions);
+
+		$storeLogEntriesFromLastSevenDays = array_filter($storeLogEntries, function ($logEntry) {
+			$performedAt = new DateTime($logEntry['performed_at']);
+			$performedAtTimestamp = $performedAt->getTimestamp();
+			$dateBeforeSevenDays = new DateTime('-7 days');
+			$dateBeforeSevenDays = $dateBeforeSevenDays->setTime(0, 0, 0, 0);
+			$dateBeforeSevenDaysTimestamp = $dateBeforeSevenDays->getTimestamp();
+
+			return $performedAtTimestamp >= $dateBeforeSevenDaysTimestamp;
+		});
+
+		$storeLogEntriesFromLastSevenDays = array_map(function ($logEntry) {
+			$correctedDate = new DateTime($logEntry['date_reference']);
+			$correctedDate->add(new \DateInterval('PT2H'));
+			$logEntry['date_reference'] = $correctedDate->format('Y-m-d H:i:s');
+
+			return $logEntry;
+		}, $storeLogEntriesFromLastSevenDays);
+
+		$extendedLogEntries = $this->extendStoreLogWithFoodsaverProfilData($storeId, $storeLogEntriesFromLastSevenDays);
+
+		return $this->handleView($this->view($extendedLogEntries, 200));
+	}
+
+	private function extendStoreLogWithFoodsaverProfilData(int $storeId, array $storeLogEntries): array
+	{
+		$storeTeam = [];
+		foreach ($this->storeGateway->getStoreTeam($storeId) as $teamMember) {
+			$foodsaverId = $teamMember['id'];
+			$storeTeam[$foodsaverId] = RestNormalization::normalizeStoreUser($teamMember);
+		}
+
+		$mergedStoreLogEntries = [];
+
+		foreach ($storeLogEntries as $entry) {
+			$affectedFoodsaverId = $entry['affected_foodsaver_id'];
+			$performedFoodsaverId = $entry['performed_foodsaver_id'];
+
+			$affectedFoodsaverModel = $storeTeam[$affectedFoodsaverId] ?? RestNormalization::normalizeStoreUser($this->foodsaverGateway->getFoodsaver($affectedFoodsaverId));
+			$performedFoodsaverModel = null;
+
+			if (!is_null($performedFoodsaverId)) {
+				$performedFoodsaverModel = $storeTeam[$performedFoodsaverId] ?? RestNormalization::normalizeStoreUser($this->foodsaverGateway->getFoodsaver($performedFoodsaverId));
+			}
+
+			$entry['affected_foodsaver'] = $affectedFoodsaverModel;
+			$entry['performed_foodsaver'] = $performedFoodsaverModel;
+
+			$mergedStoreLogEntries[] = $entry;
+		}
+
+		return $mergedStoreLogEntries;
 	}
 
 	/**